[BE] fix: 삭제 여부에 따라 호출하도록 수정

+++ BE/src/main/java/com/nts/board/service/BoardServiceImpl.java
@@ -28,6 +28,7 @@ public class BoardServiceImpl implements BoardService {
     @Override
     public BoardResponse findBoard(long boardId) {
         Board board = boardRepository.findById(boardId).orElseThrow(() -> new BoardException(BOARD_NOT_FOUND));
+        if(board.isDeleted()) throw new BoardException(BOARD_DELETED);
         board.increaseViewCount();
         return BoardResponse.from(board);
     }
@@ -49,6 +50,7 @@ public class BoardServiceImpl implements BoardService {
     @Override
     public BoardResponse updateBoard(long boardId, BoardRequest request) {
         Board findBoard = boardRepository.findById(boardId).orElseThrow(() -> new BoardException(BOARD_NOT_FOUND));
+        if(findBoard.isDeleted()) throw new BoardException(BOARD_DELETED);
         if(!passwordEncoder.matches(request.getPassword(), findBoard.getPassword()))
             throw new BoardException(BOARD_FORBIDDEN);
         findBoard.update(request.getTitle(), request.getContent());
@@ -58,6 +60,7 @@ public class BoardServiceImpl implements BoardService {
     @Override
     public BoardResponse deleteBoard(long boardId, BoardRequest request) {
         Board findBoard = boardRepository.findById(boardId).orElseThrow(() -> new BoardException(BOARD_NOT_FOUND));
+        if(findBoard.isDeleted()) throw new BoardException(BOARD_DELETED);
         if(!passwordEncoder.matches(request.getPassword(), findBoard.getPassword()))
             throw new BoardException(BOARD_FORBIDDEN);
         findBoard.delete();
@@ -68,6 +71,7 @@ public class BoardServiceImpl implements BoardService {
     public List<BoardResponse> findBoardList() {
         List<Board> boardList = boardRepository.findAll();
         return boardList.stream()
+                .filter(board -> !board.isDeleted())
                 .map(BoardResponse::from)
                 .collect(Collectors.toList());
     }
